requirements.txt
Django==3.0.2
djangorestframework==3.11.0
requests==2.22.0
psycopg2==2.8.4
numpy==1.18.1
pandas==0.25.3
google-cloud-tasks==1.3.0
google-python-cloud-debugger